requirements.txt
mesa==3.3.0
pygame-ce==2.5.6
numpy==2.3.4
scipy==1.16.2
pyyaml==6.0.3
matplotlib==3.10.7
seaborn==0.13.2